refacot(app-mobile): agrega async y modulariza calculoEdad
