+++ core/mpi/controller/paciente.ts
@@ -1088,13 +1088,60 @@ export async function buscarRelaciones(dni) {
 }
 
 export function calcularEdad(fecha) {
-    let hoy = new Date();
-    let cumpleanos = new Date(fecha);
-    let edad = hoy.getFullYear() - cumpleanos.getFullYear();
-    let m = hoy.getMonth() - cumpleanos.getMonth();
-
-    if (m < 0 || (m === 0 && hoy.getDate() < cumpleanos.getDate())) {
-        edad--;
+    let edad = null;
+    if (fecha) {
+        const birthDate = new Date(fecha);
+        const currentDate = new Date();
+        let years = (currentDate.getFullYear() - birthDate.getFullYear());
+        if (currentDate.getMonth() < birthDate.getMonth() ||
+            currentDate.getMonth() === birthDate.getMonth() && currentDate.getDate() < birthDate.getDate()) {
+            years--;
+        }
+        edad = years;
     }
+    return edad;
+}
+
+export function edadReal(fecha) {
+    // Calcula Edad de una persona (Redondea -- 30.5 años = 30 años)
+    let edad: Object;
+    let fechaNac: any;
+    const fechaActual: Date = new Date();
+    let fechaAct: any;
+    let difAnios: any;
+    let difDias: any;
+    let difMeses: any;
+    let difHs: any;
+
+    fechaNac = moment(fecha, 'YYYY-MM-DD HH:mm:ss');
+    fechaAct = moment(fechaActual, 'YYYY-MM-DD HH:mm:ss');
+    difDias = fechaAct.diff(fechaNac, 'd'); // Diferencia en días
+    difAnios = Math.floor(difDias / 365.25);
+    difMeses = Math.floor(difDias / 30.4375);
+    difHs = fechaAct.diff(fechaNac, 'h'); // Diferencia en horas
+
+
+    if (difAnios !== 0) {
+        edad = {
+            valor: difAnios,
+            unidad: 'Años'
+        };
+    } else if (difMeses !== 0) {
+        edad = {
+            valor: difMeses,
+            unidad: 'Meses'
+        };
+    } else if (difDias !== 0) {
+        edad = {
+            valor: difDias,
+            unidad: 'Días'
+        };
+    } else if (difHs !== 0) {
+        edad = {
+            valor: difHs,
+            unidad: 'Horas'
+        };
+    }
+
     return edad;
 }
